push to rubygems after build & push of OCI images
For the v1.5.0 release, the push to rubygems returned an API error even
though the push was successful. As rubygems does not allow repushing the
same version, it was not possible to rerun the workflow to build + push
the OCI images.  By reordering the sequence of steps so the push to
rubygems in the last step, this will hopefully allow for a complete
release process even though the gha job will have a failure status.